requirements.txt
flask==2.3.2
scapy==2.5.0
pillow==12.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==2.3.7
